deployment/lambdas/requirements.txt
boto3>=1.28.0
botocore>=1.31.0
pillow>=10.0.0
pdf2image>=1.16.0
